fix: missing await in Queue.put (#421)
* fix: missing await in Queue.put

* fix: exhaust_iterators

* chore: `black .`

* chore: `black .`

* feat(test): test counter lock

* chore: `black .`

* feat(test): more test semaphore tests

* fix(test): rethink failing test

* chore: `black .`

* fix: compile error

---------

+++ a_sync/primitives/locks/counter.pxd
@@ -4,9 +4,9 @@ from a_sync.primitives.locks.event cimport CythonEvent as Event
 
 cdef class CounterLock(_DebugDaemonMixin):
     cdef char* __name
-    cdef int _value
-    cdef dict[int, Event] _events
-    cpdef bint is_ready(self, int v)
-    cdef bint c_is_ready(self, int v)
-    cpdef void set(self, int value)
-    cdef void c_set(self, int value)
+    cdef long long _value
+    cdef dict[long long, Event] _events
+    cpdef bint is_ready(self, long long v)
+    cdef bint c_is_ready(self, long long v)
+    cpdef void set(self, long long value)
+    cdef void c_set(self, long long value)
